requirements.txt
anonymeter==0.0.1
autopep8==2.0.1
cufflinks==0.17.3
fire==0.5.0
pandas==1.5.3
sdmetrics==0.9.1
seaborn==0.12.2
sdv==0.18.0
gretel-client==0.15.14
ydata-sdk==0.6.0
